Move JsonTextBox and JsonErrorInfoExtensions to AppTemplate.
