+++ Sandra.UI/Localizers.cs
@@ -180,34 +180,4 @@ namespace Sandra.UI
             };
         }
     }
-
-    public static class JsonErrorInfoExtensions
-    {
-        public static LocalizedStringKey GetLocalizedStringKey(JsonErrorCode jsonErrorCode)
-        {
-            const string UnspecifiedMessage = "Unspecified error";
-
-            if (jsonErrorCode == JsonErrorCode.Unspecified)
-            {
-                return LocalizedStringKey.Unlocalizable(UnspecifiedMessage);
-            }
-            else
-            {
-                return new LocalizedStringKey($"JsonError{jsonErrorCode}");
-            }
-        }
-
-        /// <summary>
-        /// Gets the formatted and localized error message of a <see cref="JsonErrorInfo"/>.
-        /// </summary>
-        public static string Message(this JsonErrorInfo jsonErrorInfo, Localizer localizer)
-        {
-            if (jsonErrorInfo is PTypeError typeError)
-            {
-                return typeError.GetLocalizedMessage(localizer);
-            }
-
-            return localizer.Localize(GetLocalizedStringKey(jsonErrorInfo.ErrorCode), jsonErrorInfo.Parameters);
-        }
-    }
 }
